fix invalid submission id
