+++ src/app/challenge/[id]/submissions/[submissionId]/page.tsx
@@ -2,18 +2,21 @@ import { notFound } from 'next/navigation';
 import { getServerAuthSession } from '~/server/auth';
 
 import { Submissions } from '~/components/challenge/submissions';
-import { getChallengeSubmissions } from '../page';
+import { getChallengeSubmissions } from '../getChallengeSubmissions';
 interface Props {
   params: {
     id: string;
+    submissionId: string;
   };
 }
 
-export default async function SubmissionPage({ params: { id } }: Props) {
+export default async function SubmissionPage({ params: { id, submissionId } }: Props) {
   const session = await getServerAuthSession();
   const submissions = await getChallengeSubmissions(session?.user.id ?? '', id);
 
-  if (!submissions) {
+  const isValidSubmission = submissions?.some((submission) => submission.id === +submissionId);
+
+  if (!isValidSubmission) {
     return notFound();
   }
