+++ src/app/challenge/[id]/submissions/getChallengeSubmissions.ts
@@ -1,0 +1,16 @@
+import { cache } from 'react';
+import { prisma } from '~/server/db';
+
+export type ChallengeSubmissions = NonNullable<Awaited<ReturnType<typeof getChallengeSubmissions>>>;
+export const getChallengeSubmissions = cache(async (userId: string, challengeId: string) => {
+  const solutions = await prisma.submission.findMany({
+    where: { challengeId: +challengeId, userId },
+    orderBy: [
+      {
+        createdAt: 'desc',
+      },
+    ],
+  });
+
+  return solutions;
+});
